Fix java.net.SocketException exception in CouchbaseContainer.stop()
Motivation
==========
When stop() was called without a prior call to
createCouchbaseEnvironment(), it would throw a SocketException
because stopCluster() would end up trying to initialize the cluster
at the same time the container is being shut down.

Modifications
=============
Call stopCluster() before stopping the container, instead of in
parallel.

Result
======
stopCluster() might still call createCouchbaseEnvironment(),
but the call succeeds and the environment is immediately cleaned up.